Split tmux presentation tests by responsibility
The first tmux workflow split left presentation coverage above 3,000 lines.
That file still mixed retained-client behavior with preview capture, confirmed
kills, and explicit session creation, which made focused changes harder to
review.

Give each behavior group its own source file while preserving the existing
serialized suite, test identities, fixtures, and scheduling. Production code
and access boundaries remain unchanged.

- Make rapid sidebar toggle coverage independent of animation duration.
